fix: an ut error
