+++ packages/ui/package.json
@@ -81,6 +81,7 @@
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.9",
     "@types/react-flatpickr": "3.8.8",
+    "@types/testing-library__jest-dom": "5.14.5",
     "autoprefixer": "10.4.13",
     "babel-loader": "8.2.5",
     "chromatic": "6.11.4",
