+++ packages/ui/src/__tests__/jest.setup.ts
@@ -26,7 +26,3 @@ jest.mock('next/dynamic', () => (func: () => Promise<any>) => {
   DynamicComponent.preload = jest.fn();
   return DynamicComponent;
 });
-
-jest.mock('rehype-pretty-code', () => () => null);
-jest.mock('rehype-autolink-headings', () => () => null);
-jest.mock('rehype-slug', () => () => null);
