+++ packages/ui/src/blocks/notification-hub/notification-hub.tsx
@@ -56,13 +56,14 @@ export function NotificationHub(): JSX.Element {
                   index={index}
                   length={data.length}
                   onClickCapture={async (messageId) => {
-                    if (!!!process.env.NEXT_PUBLIC_MAINTENANCE_MODE) {
+                    if (!!process.env.NEXT_PUBLIC_MAINTENANCE_MODE) {
                       return;
                     }
                     await readANotification({ messageId });
                     await refechMessages();
                   }}
                   onClickDelete={async (messageId) => {
+                    console.log('Clicking delete');
                     await deleteNotificationMessage({ messageId });
                     refechMessages();
                   }}
