+++ packages/ui/tsconfig.json
@@ -3,7 +3,7 @@
   "extends": "@chirpy-dev/tsconfigs/base.json",
   "compilerOptions": {
     "baseUrl": ".",
-    "types": ["jest", "node"]
+    "types": ["jest", "node", "@testing-library/jest-dom"]
   },
   "exclude": ["node_modules"],
   "include": ["typings/**/*.d.ts", "src/**/*.ts", "src/**/*.tsx"]
